build!: switch from rollup to esbuild to build
BREAKING CHANGE: UMD is no longer exported.

In order to reduce complexity and simplify the build process,
we have switched from rollup to esbuild and only export CJS and ESM versions.

This means that the UMD build is no longer exported.
