+++ package.json
@@ -2,16 +2,26 @@
   "name": "react-currency-input-field",
   "version": "3.8.0",
   "description": "React <input/> component for formatting currency and numbers.",
+  "sideEffects": false,
   "files": [
     "dist/**/*",
     "src/**/*"
   ],
-  "main": "dist/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.js"
+    }
+  },
+  "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
   "types": "dist/index.d.ts",
   "homepage": "https://github.com/cchanxzy/react-currency-input-field",
   "scripts": {
-    "build": "rm -rf dist && tsc && rollup -c",
+    "build:types": "tsc",
+    "build:js": "node esbuild.mjs",
+    "build": "rm -rf dist && pnpm build:js && pnpm build:types",
     "start": "parcel src/examples/index.html --open",
     "test": "LANG=en_GB jest",
     "test-ci": "LANG=en_GB.UTF-8 cross-env NODE_ICU_DATA=node_modules/full-icu jest",
@@ -48,9 +58,6 @@
     "@commitlint/cli": "^18.4.3",
     "@commitlint/config-conventional": "^18.4.3",
     "@playwright/test": "^1.40.1",
-    "@rollup/plugin-commonjs": "^19.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@rollup/plugin-typescript": "^8.2.1",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
     "@testing-library/dom": "^9.3.3",
@@ -65,6 +72,7 @@
     "@typescript-eslint/parser": "^6.16.0",
     "codecov": "^3.8.3",
     "cross-env": "^7.0.3",
+    "esbuild": "^0.24.0",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.2",
@@ -72,6 +80,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "full-icu": "^1.3.4",
     "gh-pages": "^5.0.0",
+    "glob": "^11.0.0",
     "husky": "^8.0.3",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
@@ -81,7 +90,6 @@
     "process": "^0.11.10",
     "react": "^18.2.0",
     "react-dom": "18.2.0",
-    "rollup": "^2.52.7",
     "semantic-release": "^24.2.0",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.2",
