+++ src/components/utils/formatValue.ts
@@ -1,58 +1,6 @@
-import { IntlConfig } from '../CurrencyInputProps';
 import { escapeRegExp } from './escapeRegExp';
 import { getSuffix } from './getSuffix';
-
-export type FormatValueOptions = {
-  /**
-   * Value to format
-   */
-  value: string | undefined;
-
-  /**
-   * Decimal separator
-   *
-   * Default = '.'
-   */
-  decimalSeparator?: string;
-
-  /**
-   * Group separator
-   *
-   * Default = ','
-   */
-  groupSeparator?: string;
-
-  /**
-   * Turn off separators
-   *
-   * This will override Group separators
-   *
-   * Default = false
-   */
-  disableGroupSeparators?: boolean;
-
-  /**
-   * Intl locale currency config
-   */
-  intlConfig?: IntlConfig;
-
-  /**
-   * Specify decimal scale for padding/trimming
-   *
-   * Eg. 1.5 -> 1.50 or 1.234 -> 1.23
-   */
-  decimalScale?: number;
-
-  /**
-   * Prefix
-   */
-  prefix?: string;
-
-  /**
-   * Suffix
-   */
-  suffix?: string;
-};
+import { FormatValueOptions } from './formatValue.types';
 
 /**
  * Format value with decimal separator, group separator and prefix
