+++ src/index.ts
@@ -1,6 +1,12 @@
 import CurrencyInput from './components/CurrencyInput';
 
-export { CurrencyInputProps, CurrencyInputOnChangeValues } from './components/CurrencyInputProps';
-export default CurrencyInput;
+export type {
+  CurrencyInputProps,
+  CurrencyInputOnChangeValues,
+} from './components/CurrencyInputProps';
+
+export { CurrencyInput } from './components/CurrencyInput';
 export { formatValue } from './components/utils/formatValue';
 export { cleanValue } from './components/utils/cleanValue';
+
+export default CurrencyInput;
